requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
praw==7.6.0
prawcore==2.3.0
PyYAML==6.0
requests==2.28.1
update-checker==0.18.0
urllib3==1.26.9
websocket-client==1.3.3
